Refactor file dialog service to fix layer breaker
Fixes #81161
