+++ src/vs/workbench/services/dialogs/electron-browser/fileDialogService.ts
@@ -1,0 +1,162 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { IWindowService, OpenDialogOptions, SaveDialogOptions } from 'vs/platform/windows/common/windows';
+import { IPickAndOpenOptions, ISaveDialogOptions, IOpenDialogOptions, IFileDialogService } from 'vs/platform/dialogs/common/dialogs';
+import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
+import { IHistoryService } from 'vs/workbench/services/history/common/history';
+import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
+import { URI } from 'vs/base/common/uri';
+import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
+import { IFileService } from 'vs/platform/files/common/files';
+import { IOpenerService } from 'vs/platform/opener/common/opener';
+import { IElectronService } from 'vs/platform/electron/node/electron';
+import { AbstractFileDialogService } from 'vs/workbench/services/dialogs/browser/abstractFileDialogService';
+
+export class FileDialogService extends AbstractFileDialogService implements IFileDialogService {
+
+	_serviceBrand: undefined;
+
+	constructor(
+		@IWindowService windowService: IWindowService,
+		@IWorkspaceContextService contextService: IWorkspaceContextService,
+		@IHistoryService historyService: IHistoryService,
+		@IWorkbenchEnvironmentService environmentService: IWorkbenchEnvironmentService,
+		@IInstantiationService instantiationService: IInstantiationService,
+		@IConfigurationService configurationService: IConfigurationService,
+		@IFileService fileService: IFileService,
+		@IOpenerService openerService: IOpenerService,
+		@IElectronService private readonly electronService: IElectronService
+	) { super(windowService, contextService, historyService, environmentService, instantiationService, configurationService, fileService, openerService); }
+
+	async pickFileFolderAndOpen(options: IPickAndOpenOptions): Promise<any> {
+		const schema = this.getFileSystemSchema(options);
+
+		if (!options.defaultUri) {
+			options.defaultUri = this.defaultFilePath(schema);
+		}
+
+		const shouldUseSimplified = this.shouldUseSimplified(schema);
+		if (shouldUseSimplified.useSimplified) {
+			return this.pickFileFolderAndOpenSimplified(schema, options, shouldUseSimplified.isSetting);
+		}
+		return this.electronService.pickFileFolderAndOpen(this.toNativeOpenDialogOptions(options));
+	}
+
+	async pickFileAndOpen(options: IPickAndOpenOptions): Promise<any> {
+		const schema = this.getFileSystemSchema(options);
+
+		if (!options.defaultUri) {
+			options.defaultUri = this.defaultFilePath(schema);
+		}
+
+		const shouldUseSimplified = this.shouldUseSimplified(schema);
+		if (shouldUseSimplified.useSimplified) {
+			return this.pickFileAndOpenSimplified(schema, options, shouldUseSimplified.isSetting);
+		}
+		return this.electronService.pickFileAndOpen(this.toNativeOpenDialogOptions(options));
+	}
+
+	async pickFolderAndOpen(options: IPickAndOpenOptions): Promise<any> {
+		const schema = this.getFileSystemSchema(options);
+
+		if (!options.defaultUri) {
+			options.defaultUri = this.defaultFolderPath(schema);
+		}
+
+		if (this.shouldUseSimplified(schema).useSimplified) {
+			return this.pickFolderAndOpenSimplified(schema, options);
+		}
+		return this.electronService.pickFolderAndOpen(this.toNativeOpenDialogOptions(options));
+	}
+
+	async pickWorkspaceAndOpen(options: IPickAndOpenOptions): Promise<void> {
+		const schema = this.getFileSystemSchema(options);
+
+		if (!options.defaultUri) {
+			options.defaultUri = this.defaultWorkspacePath(schema);
+		}
+
+		if (this.shouldUseSimplified(schema).useSimplified) {
+			return this.pickWorkspaceAndOpenSimplified(schema, options);
+		}
+		return this.electronService.pickWorkspaceAndOpen(this.toNativeOpenDialogOptions(options));
+	}
+
+	async pickFileToSave(options: ISaveDialogOptions): Promise<URI | undefined> {
+		const schema = this.getFileSystemSchema(options);
+		if (this.shouldUseSimplified(schema).useSimplified) {
+			return this.pickFileToSaveSimplified(schema, options);
+		} else {
+			const result = await this.electronService.showSaveDialog(this.toNativeSaveDialogOptions(options));
+			if (result && result.filePath) {
+				return URI.file(result.filePath);
+			}
+		}
+		return;
+	}
+
+	private toNativeSaveDialogOptions(options: ISaveDialogOptions): SaveDialogOptions {
+		options.defaultUri = options.defaultUri ? URI.file(options.defaultUri.path) : undefined;
+		return {
+			defaultPath: options.defaultUri && options.defaultUri.fsPath,
+			buttonLabel: options.saveLabel,
+			filters: options.filters,
+			title: options.title
+		};
+	}
+
+	async showSaveDialog(options: ISaveDialogOptions): Promise<URI | undefined> {
+		const schema = this.getFileSystemSchema(options);
+		if (this.shouldUseSimplified(schema).useSimplified) {
+			return this.showSaveDialogSimplified(schema, options);
+		}
+
+		const result = await this.electronService.showSaveDialog(this.toNativeSaveDialogOptions(options));
+		if (result && result.filePath) {
+			return URI.file(result.filePath);
+		}
+
+		return;
+	}
+
+	async showOpenDialog(options: IOpenDialogOptions): Promise<URI[] | undefined> {
+		const schema = this.getFileSystemSchema(options);
+		if (this.shouldUseSimplified(schema).useSimplified) {
+			return this.showOpenDialogSimplified(schema, options);
+		}
+
+		const defaultUri = options.defaultUri;
+
+		const newOptions: OpenDialogOptions = {
+			title: options.title,
+			defaultPath: defaultUri && defaultUri.fsPath,
+			buttonLabel: options.openLabel,
+			filters: options.filters,
+			properties: []
+		};
+
+		newOptions.properties!.push('createDirectory');
+
+		if (options.canSelectFiles) {
+			newOptions.properties!.push('openFile');
+		}
+
+		if (options.canSelectFolders) {
+			newOptions.properties!.push('openDirectory');
+		}
+
+		if (options.canSelectMany) {
+			newOptions.properties!.push('multiSelections');
+		}
+
+		const result = await this.electronService.showOpenDialog(newOptions);
+		return result && result.filePaths ? result.filePaths.map(URI.file) : undefined;
+	}
+}
+
+registerSingleton(IFileDialogService, FileDialogService, true);
